VectordB/requirements.txt
streamlit
openai
chromadb
python-dotenv
numpy
google-search-results
beautifulsoup4
requests
